# src/main/resources/api-def/calculationApi-openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: Calculation API
    description: Calculation controller
paths:
  /calculation:
    post:
      tags:
        - Calculation API
      operationId: postCalculation
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CalculationRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResponseFull'
  /calculation/pagination:
    get:
      tags:
        - Calculation API
      operationId: getAllCalculations
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCalculationResponseFull'
components:
  schemas:
    CalculationRequest:
      required:
        - calculationType
        - values
      type: object
      properties:
        values:
          type: array
          description: Required
          items:
            type: integer
            description: Required
            format: int32
        calculationType:
          type: string
          description: Required
          enum:
            - SUM
            - SUBTRATION
            - MULTIPLICATION
            - DIVISION
    CalculationResponseFull:
      type: object
      properties:
        id:
          type: integer
          format: int64
        calculationDate:
          type: string
          format: date-time
        calculations:
          type: array
          items:
            $ref: '#/components/schemas/MathResponse'
    MathResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        valuesCalculation:
          type: array
          items:
            type: integer
            format: int32
        calculationType:
          type: string
          enum:
            - SUM
            - SUBTRATION
            - MULTIPLICATION
            - DIVISION
        result:
          type: number
          format: double
    PageCalculationResponseFull:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CalculationResponseFull'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        unsorted:
          type: boolean
        sorted:
          type: boolean
        empty:
          type: boolean